+++ game_patch/multi/server_internal.h
@@ -49,6 +49,7 @@ struct ServerAdditionalConfig
     VoteConfig vote_restart;
     VoteConfig vote_next;
     VoteConfig vote_previous;
+    VoteConfig vote_match;
     NewSpawnLogicConfig new_spawn_logic;
     int spawn_protection_duration_ms = 1500;
     std::optional<float> spawn_life;
@@ -78,11 +79,57 @@ struct ServerAdditionalConfig
     bool kill_reward_armor_super = false;
 };
 
+struct MatchInfo
+{
+    std::time_t last_match_reminder_time = 0;
+    bool pre_match_queued = false;
+    bool pre_match_active = false;
+    std::time_t pre_match_start_time = 0;
+    std::optional<float> time_limit_on_pre_match_start;
+    std::time_t last_ready_reminder_time = 0;
+    bool everyone_ready = false;
+    bool match_active = false;
+    int team_size = -1;
+    std::set<rf::Player*> ready_players_red;
+    std::set<rf::Player*> ready_players_blue;
+    std::set<rf::Player*> active_match_players;
+    std::string match_level_name;
+
+    void reset()
+    {
+        last_match_reminder_time = 0;
+        pre_match_queued = false;
+        pre_match_active = false;
+        pre_match_start_time = 0;
+        time_limit_on_pre_match_start.reset();
+        last_ready_reminder_time = 0;
+        everyone_ready = false;
+        match_active = false;
+        team_size = -1;
+        ready_players_red.clear();
+        ready_players_blue.clear();
+        active_match_players.clear();
+        match_level_name.clear();
+    }
+};
+
 extern ServerAdditionalConfig g_additional_server_config;
 extern std::string g_prev_level;
+extern MatchInfo g_match_info;
+
+
 
 void cleanup_win32_server_console();
 void handle_vote_command(std::string_view vote_name, std::string_view vote_arg, rf::Player* sender);
+std::set<rf::Player*> get_current_player_list(bool include_browsers);
+std::pair<bool, std::string> is_level_name_valid(std::string_view level_name_input);
+bool is_player_in_match(rf::Player* player);
+bool is_player_ready(rf::Player* player);
+void update_pre_match_powerups(rf::Player* player);
+void start_match();
+void cancel_match();
+void start_pre_match();
+void set_ready_status(rf::Player* player, bool is_ready);
 void server_vote_do_frame();
 void init_server_commands();
 void extend_round_time(int minutes);