+++ docs/CHANGELOG.md
@@ -54,6 +54,7 @@ Version 1.9.0 (not released yet)
 - Make game spawn players at Player Start position if level has no respawn points
 - Add respawn point selection logic settings to dedicated server config
 - Raise limit on number of multiplayer respawn points per level from 32 to 2048
+- Add match mode ready up system (controlled by `$DF Vote Match` option in dedicated server config)
 
 Version 1.8.0 (released 2022-09-17)
 -----------------------------------